Address implicit conversions

+++ include/configuration_property.tpl.hpp
@@ -207,7 +207,7 @@ template<class T> SizeType EnumConfigurationProperty<T>::cardinality() const {
 
 template<class T> List<int> EnumConfigurationProperty<T>::local_integer_values() const {
     List<int> result;
-    for (SizeType i=0; i<_values.size(); ++i) result.push_back(i);
+    for (SizeType i=0; i<_values.size(); ++i) result.push_back(static_cast<int>(i));
     return result;
 }
 
@@ -440,7 +440,7 @@ template<class T> SizeType InterfaceListConfigurationProperty<T>::cardinality() 
 
 template<class T> List<int> InterfaceListConfigurationProperty<T>::local_integer_values() const {
     List<int> result;
-    for (SizeType i=0; i<_values.size(); ++i) result.push_back(i);
+    for (SizeType i=0; i<_values.size(); ++i) result.push_back(static_cast<int>(i));
     return result;
 }
